requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.115.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
